rollback to old stake implementation
